Move ClientUtilsTest.FormatPropertyDescriptions to configuration module
