+++ java/clients/src/main/java/sleeper/util/ClientUtils.java
@@ -17,21 +17,16 @@ package sleeper.util;
 
 import com.amazonaws.services.s3.AmazonS3;
 import com.amazonaws.services.s3.AmazonS3ClientBuilder;
-import org.apache.commons.lang.WordUtils;
 
 import sleeper.configuration.properties.InstanceProperties;
-import sleeper.configuration.properties.PropertyGroup;
-import sleeper.configuration.properties.SleeperProperty;
 
 import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Comparator;
 import java.util.List;
 import java.util.Optional;
-import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
 public class ClientUtils {
@@ -102,20 +97,6 @@ public class ClientUtils {
             parts.add(str.substring(i, i + 3));
         }
         return String.join(",", parts);
-    }
-
-    public static String formatDescription(SleeperProperty property) {
-        return formatString(property.getDescription());
-    }
-
-    public static String formatDescription(PropertyGroup group) {
-        return formatString(group.getDescription());
-    }
-
-    public static String formatString(String str) {
-        return Arrays.stream(str.split("\n")).
-                map(line -> "# " + WordUtils.wrap(line, 100).replace("\n", "\n# "))
-                .collect(Collectors.joining("\n"));
     }
 
     public static void clearDirectory(Path tempDir) throws IOException {
